Moved non-view functions to utils.py
