+++ server/flaskr/views.py
@@ -6,11 +6,11 @@ import os
 import datetime
 import json
 import sys
-import tarfile
-import zipfile
 
-# sys.path.insert(0, '../../')
+# Custom imports
 from GOF_templates import render
+
+from server.flaskr import utils
 
 app.secret_key = 'secretkeyhereplease'
 
@@ -83,33 +83,13 @@ def codeCreate():
 
 @app.route("/codeDownload/<path:filename>/<path:patternType>/<path:fileType>",methods=["GET","POST"])
 def codeDownload(filename,patternType,fileType):
-    # START: Check if a downloads folder exists
-    dirname = os.path.dirname(os.path.abspath(__file__))
-    try: 
-        os.makedirs(os.path.join(dirname,'downloads'))
-        print("\n\n\nDownloads folder Created!\n\n\n")
-    except(FileExistsError):
-        print("\n\n\nDownloads folder already exists!\n\n\n")
-        pass
-    # END: Check if a downloads folder exists 
-    makeCompressedfile(outputFilename=os.path.join(app.config['CODE_DOWNLOAD_FOLDER'],filename),
+    
+    utils.checkAndCreateDownloadsFolder()
+
+    utils.makeCompressedfile(outputFilename=os.path.join(app.config['CODE_DOWNLOAD_FOLDER'],filename),
                        sourceDir=os.path.join("./GOF_templates/templates/output/",patternType),
                        fileType=fileType)
 
     print("Path:",os.path.join(app.config['USER_DOWNLOAD_FOLDER'],filename))
     return send_from_directory(app.config['USER_DOWNLOAD_FOLDER'],filename,as_attachment=True)
-
-def makeCompressedfile(outputFilename, sourceDir, fileType):
-    if(fileType.startswith(".tar")):
-        extension = fileType.split(".")[2]
-        with tarfile.open(outputFilename, "w:"+extension) as tar:
-            tar.add(sourceDir, arcname=os.path.basename(sourceDir))
-    
-    else:
-        print("outputFilename:",outputFilename)
-        with zipfile.ZipFile(outputFilename, 'w', zipfile.ZIP_DEFLATED) as ziph:
-            for root, dirs, files in os.walk(sourceDir):
-                for file in files:
-                    ziph.write(os.path.join(root, file))
-
